+++ src/main/java/org/realityforge/ssf/InMemorySessionManager.java
@@ -15,6 +15,7 @@ public abstract class InMemorySessionManager<T extends SessionInfo>
 {
   private final Map<String, T> _sessions =
     Collections.synchronizedMap( new HashMap<String, T>() );
+  private final Map<String, T> _roSessions = Collections.unmodifiableMap( _sessions );
 
   /**
    * {@inheritDoc}
@@ -73,10 +74,15 @@ public abstract class InMemorySessionManager<T extends SessionInfo>
     return sessionInfo;
   }
 
+  /**
+   * Return an unmodifiable map containing the set of sessions.
+   *
+   * @return an unmodifiable map containing the set of sessions.
+   */
   @Nonnull
   protected final Map<String, T> getSessions()
   {
-    return _sessions;
+    return _roSessions;
   }
 
   /**