+++ src/test/java/org/realityforge/ssf/SimpleSessionManagerTest.java
@@ -26,7 +26,7 @@ public class SimpleSessionManagerTest
     Thread.sleep( 1 );
 
     // Make sure we can also get it thorugh the map interface
-    assertEquals( sm.getSessions().get(  sessionInfo.getSessionID() ), sessionInfo );
+    assertEquals( sm.getSessions().get( sessionInfo.getSessionID() ), sessionInfo );
 
     // The next line should update the last accessed time too!
     assertEquals( sm.getSession( sessionInfo.getSessionID() ), sessionInfo );